fix npc distance db

+++ app/model/database/DatabaseConfig.js
@@ -98,7 +98,7 @@ CREATE TABLE IF NOT EXISTS VOMSSymptoms (
 CREATE TABLE IF NOT EXISTS VOMSNPCDistance (
     vomsNPCDistance_id INTEGER PRIMARY KEY,
     report_id INTEGER REFERENCES IncidentReport(report_id),
-    distance FLOAT,
+    distance FLOAT(20),
 );
 `,
 ];
